# nova-backend/swagger.yaml
openapi: 3.0.0
info:
  title: Nova API
  description: Nova School Administration API Documentation
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local Server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /admin/login:
    post:
      summary: Admin Login
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  username:
                    type: string
        '401':
          description: Invalid credentials

  /apply:
    post:
      summary: Submit a new application
      tags: [Public]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                parentName:
                  type: string
                parentPhone:
                  type: string
                childName:
                  type: string
                childDob:
                  type: string
                grade:
                  type: string
      responses:
        '201':
          description: Application submitted successfully

  /public/news:
    get:
      summary: Get all news (Paginated)
      tags: [Public]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
          description: Items per page
      responses:
        '200':
          description: A list of news

  /admin/applications:
    get:
      summary: Get all applications (Paginated)
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
          description: Items per page
      responses:
        '200':
          description: A list of applications
        '401':
          description: Unauthorized